requirements.txt
flask
torch>=2.1.0
Pillow
ultralytics>=8.1.28
gunicorn
requests
